src/requirements.txt
asgiref==3.2.10
Django==2.2.2
django-phonenumber-field==5.0.0
phonenumbers==8.12.10
Pillow==8.0.1
pytz==2020.1
sqlparse==0.3.1
